Further major refactoring
Refactoring of frames to separate raw frames from AMQP and SASL frames.
Various other changes.

+++ Core/Msg.Core/Transport/Frames/FrameHeader.cs
@@ -1,8 +1,10 @@
+using System;
+
 namespace Msg.Core.Transport.Frames
 {
     public class FrameHeader
     {
-        public FrameHeader (uint size, uint dataOffset, FrameHeaderType type, uint channelId)
+        public FrameHeader(uint size, byte dataOffset, FrameHeaderType type, ushort channelId)
         {
             Size = size;
             DataOffset = dataOffset;
@@ -12,15 +14,15 @@ namespace Msg.Core.Transport.Frames
 
         public uint Size { get; private set; }
 
-        public uint DataOffset { get; private set; }
+        public byte DataOffset { get; private set; }
 
         public FrameHeaderType Type { get; private set; }
 
-        public uint ChannelId { get; private set; }
+        public ushort ChannelId { get; private set; }
 
-        public override string ToString ()
+        public override string ToString()
         {
-            return string.Format ("[FrameHeader: Size={0}, DataOffset={1}, Type={2}, ChannelId={3}]", Size, DataOffset, Type, ChannelId);
+            return string.Format("[FrameHeader: Size={0}, DataOffset={1}, Type={2}, ChannelId={3}]", Size, DataOffset, Type, ChannelId);
         }
     }
 }
